Move package level documentation to doc.go for annoying reasons
The annoying reasons are 'limitations in Go tooling'. Various tools
such as godoc and goimports do not process files that are not
applicable to the current architecture; this means that they see very
little or nothing if you run them on this package on non-Solaris
architectures. This has various bad effects.

To get around the worst of the problems we need a generic Go file in
here so that tools can discover the package name. Once we have that
generic file, we might as well put the package level documentation in it
to increase the number of tools and people who can at least see that
level of documentation.

(They probably won't see API level documentation but I refuse to try to
fix that; it'd be an awful mess.)