Document/clarify when stat data is updated; comment fixes; minor code tuneups
This includes an API change, because I realized that Named must carry a
Snaptime field since its KStat may have been Refresh'd since the Named
was taken. This is especially so because .Lookup() explicitly refreshes
the KStat every time it's called.

As a result of looking into all of this, I now understand when data is
and isn't updated (and with it ks_snaptime). There are more tests and
more documentation for this area because it's tricky.