Add support for IO kstats
What you get is an exact copy of the kstat_io_t structure, with no
extra trimmings. As part of adding this, there are now Go types for
the various RawStat things that I intend to support, since all of
them were generated through cgo and I might as well do that all at
once.

There are documentation touchups and so on as well.