Correctly copy C 'char field[SIZE];' fields to Go strings
We need an equivalent of C strndup() that copies strings that are at most
N characters long but potentially null-terminated before then. Neither
of the standard C.GoString* calls do this; C.GoString() is not length
bounded, and C.GoStringN() copies even after the null termination if
you have a short field value (which is the usual case).

I've added some tests to look for problems in this area in all of
kstat Module/Name/Class fields, named kstat Name field, and named kstat
CharData. The last test assumes that CPU 0 is online, which may result
in false negatives; unfortunately I don't know of a better CharData
field to test.