Improve and minimize my strndup() function
As Ian Lance Taylor taught me in my Github report about the general
issue, this is really 'C.GoStringN(p, C.strnlen(p, len))'. So now we
just do that. strnlen() is supported on Solaris and this is Solaris
specific code, so I don't have to worry about whether or not it exists
on all platforms.